move fixTables out of html into xml (new)
