+++ src/index.js
@@ -3,3 +3,4 @@ export * from './html';
 export * from './eids';
 export * from './enrichments';
 export * from './ranges';
+export * from './xml';
